LayoutTests/media/session/content-session-focus-expected.txt

Playing a media element belonging to a Content media session should pause other playing Content media elements.


EVENT(canplaythrough)
EXPECTED (video1.paused == 'true') OK
Begin playing first video.
EVENT(playing)
First video began playing.
EXPECTED (video1.paused == 'false') OK

Setting up second video.
EVENT(canplaythrough)
EXPECTED (video1.paused == 'false') OK
EXPECTED (video2.paused == 'true') OK
Begin playing second video.
EVENT(playing)
Second video began playing.
EXPECTED (video2.paused == 'false') OK
EVENT(pause)
EXPECTED (video1.paused == 'true') OK
END OF TEST
